requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
django-cors-headers>=4.3
Pillow>=10.0
channels>=4.0
channels-redis>=4.1
